rename: 지역 필터컴포넌트 이름 변경
